+++ src/functions.rs
@@ -1,3 +1,4 @@
+pub mod explog;
 pub mod fraction;
 pub mod power;
 pub mod sign;