+++ src/functions.rs
@@ -2,3 +2,4 @@ pub mod explog;
 pub mod fraction;
 pub mod power;
 pub mod sign;
+pub mod trigonometry;